# docs/swagger.yaml
basePath: /api
definitions:
...
  models.Mahasiswa:
    properties:
      angkatan:
        type: string
      dosen_wali_id:
        type: integer
      id:
        type: string
      nama_mhs:
        type: string
      nim:
        type: string
      prodi:
        type: string
      user_id:
        type: string
    type: object
  models.User:
    properties:
      id:
        type: string
      is_active:
        type: boolean
      nama:
        type: string
      password:
        type: string
      role_id:
        type: string
      username:
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: arum@mail.com
    name: Arum
  description: API Sistem Prestasi Mahasiswa
  title: UAS Backend API
  version: "1.0"
paths:
...
  /v1/achievements/{id}/reject:
    post:
      description: Reject an achievement submission
      parameters:
      - description: Achievement ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Reject achievement
      tags:
      - Achievement
  /v1/achievements/{id}/verify:
    post:
      description: Verify an achievement submission
      parameters:
      - description: Achievement ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Verify achievement
      tags:
      - Achievement
...
  /v1/mahasiswa:
    post:
      consumes:
      - multipart/form-data
      description: Membuat data mahasiswa baru
      parameters:
      - description: User ID
        in: formData
        name: user_id
        required: true
        type: string
      - description: NIM Mahasiswa
        in: formData
        name: nim
        required: true
        type: string
      - description: Nama Mahasiswa
        in: formData
        name: nama_mhs
        required: true
        type: string
      - description: Angkatan
        in: formData
        name: angkatan
        required: true
        type: integer
      - description: Program Studi
        in: formData
        name: prodi
        required: true
        type: string
      - description: ID Dosen Wali
        in: formData
        name: dosen_wali_id
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Mahasiswa'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create mahasiswa
      tags:
      - Mahasiswa
  /v1/mahasiswa/{id}:
    delete:
      description: Menghapus data mahasiswa berdasarkan ID
      parameters:
      - description: Mahasiswa ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Delete mahasiswa
      tags:
      - Mahasiswa
    get:
      description: Mengambil detail mahasiswa berdasarkan ID
      parameters:
      - description: Mahasiswa ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Mahasiswa'
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get detail mahasiswa
      tags:
      - Mahasiswa
    put:
      consumes:
      - application/json
      description: Mengupdate data mahasiswa berdasarkan ID
      parameters:
      - description: Mahasiswa ID
        in: path
        name: id
        required: true
        type: string
      - description: Mahasiswa Data
        in: body
        name: mahasiswa
        required: true
        schema:
          $ref: '#/definitions/models.Mahasiswa'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update mahasiswa
      tags:
      - Mahasiswa
  /v1/mahasiswa/advisor/{dosen_wali_id}:
    get:
      description: Mengambil daftar mahasiswa berdasarkan dosen wali
      parameters:
      - description: Dosen Wali ID
        in: path
        name: Dosen_wali_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Mahasiswa'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get mahasiswa bimbingan dosen
      tags:
      - Mahasiswa
  /v1/mahasiswa/user/{user_id}:
    get:
      description: Mengambil data mahasiswa berdasarkan user ID
      parameters:
      - description: User ID
        in: path
        name: user_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Mahasiswa'
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get mahasiswa by user ID
      tags:
      - Mahasiswa
  /v1/reports/statistics:
    get:
      description: Retrieve various statistics
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get statistics
      tags:
      - Report
  /v1/users/:
    get:
      description: Retrieve a list of all users
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.User'
            type: array
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get all users
      tags:
      - User
  /v1/users/{id}:
    delete:
      description: Delete user by ID
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Delete user
      tags:
      - User
    put:
      consumes:
      - application/json
      description: Update user information by ID
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      - description: User Data
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/models.User'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update user
      tags:
      - User
schemes:
- http
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"