# handlers/report_handler.go
package handlers

import (
	"net/http"

	"github.com/gin-gonic/gin"
)

// GetStatistics godoc
// @Summary Get statistics
// @Description Retrieve various statistics
// @Tags Report
// @Security BearerAuth
// @Produce json
// @Success 200 {object} map[string]interface{}
// @Failure 500 {object} map[string]string
// @Router /v1/reports/statistics [get]
// GetStatistics handles /reports/statistics
func GetStatistics(c *gin.Context) {
	c.JSON(http.StatusOK, gin.H{
		"message": "Statistics data (placeholder)",
	})
}